First attempt to implement FP16/FP32 switch

+++ nsff_exp/render_utils.py
@@ -322,7 +322,7 @@ def render_sm(img_idx, chain_bwd, chain_5frames,
         # special case to visualize effect of viewdirs
         rays_o, rays_d = get_rays(H, W, focal, c2w_staticcam)
     viewdirs = viewdirs / torch.norm(viewdirs, dim=-1, keepdim=True)
-    viewdirs = torch.reshape(viewdirs, [-1,3]).float()
+    viewdirs = torch.reshape(viewdirs, [-1,3])
 
     sh = rays_d.shape # [..., 3]
 
@@ -331,8 +331,8 @@ def render_sm(img_idx, chain_bwd, chain_5frames,
         rays_o, rays_d = ndc_rays(H, W, focal, 1., rays_o, rays_d)
 
     # Create ray batch
-    rays_o = torch.reshape(rays_o, [-1,3]).float()
-    rays_d = torch.reshape(rays_d, [-1,3]).float()
+    rays_o = torch.reshape(rays_o, [-1,3])
+    rays_d = torch.reshape(rays_d, [-1,3])
 
     near, far = near * torch.ones_like(rays_d[...,:1]), far * torch.ones_like(rays_d[...,:1])
     rays = torch.cat([rays_o, rays_d, near, far], -1)
@@ -584,7 +584,7 @@ def render(img_idx, chain_bwd, chain_5frames,
         # special case to visualize effect of viewdirs
         rays_o, rays_d = get_rays(H, W, focal, c2w_staticcam)
     viewdirs = viewdirs / torch.norm(viewdirs, dim=-1, keepdim=True)
-    viewdirs = torch.reshape(viewdirs, [-1,3]).float()
+    viewdirs = torch.reshape(viewdirs, [-1,3])
 
     sh = rays_d.shape # [..., 3]
 
@@ -593,8 +593,8 @@ def render(img_idx, chain_bwd, chain_5frames,
         rays_o, rays_d = ndc_rays(H, W, focal, 1., rays_o, rays_d)
 
     # Create ray batch
-    rays_o = torch.reshape(rays_o, [-1,3]).float()
-    rays_d = torch.reshape(rays_d, [-1,3]).float()
+    rays_o = torch.reshape(rays_o, [-1,3])
+    rays_d = torch.reshape(rays_d, [-1,3])
 
     near, far = near * torch.ones_like(rays_d[...,:1]), far * torch.ones_like(rays_d[...,:1])
     rays = torch.cat([rays_o, rays_d, near, far], -1)
